move more transformation code out to the correct file. Mostly so we're not calling methods defined in the handler in the transformation tests. Also to simplify the handler
